# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API do AmoConecta
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /safety-mode:
    get:
      operationId: getSafetyMode
      tags: [safety]
      summary: Retorna o estado seguro da operação de envios
      description: Não expõe os endereços configurados na allowlist.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Estado do modo de segurança
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafetyMode"
        "401":
          description: Usuário não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /campaign-defaults:
    get:
      operationId: getCampaignDefaults
      tags: [campaigns]
      summary: Retorna os padrões seguros para novas campanhas
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Padrões da campanha
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDefaults"
        "401":
          description: Usuário não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Configuração de remetente ausente ou inválida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /email-branding:
    get:
      operationId: getEmailBranding
      tags: [settings]
      summary: Retorna a identidade visual dos e-mails
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Configuração global atual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailBranding"
        "401":
          description: Usuário não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateEmailBranding
      tags: [settings]
      summary: Atualiza a cor global dos botões de e-mail
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailBrandingUpdate"
      responses:
        "200":
          description: Configuração atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailBranding"
        "400":
          description: Cor inválida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Usuário não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /engagement/summary:
    get:
      operationId: getEngagementSummary
      tags: [engagement]
      summary: Retorna a quantidade atual de contatos desengajados
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Resumo de engajamento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngagementSummary"
        "401":
          description: Usuário não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /bi-exports:
    get:
      operationId: listBiExports
      tags: [engagement]
      summary: Lista exportações BI recentes
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Exportações recentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BiExportJobsResponse"
        "401":
          description: Usuário não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createBiExport
      tags: [engagement]
      summary: Solicita uma exportação BI assíncrona
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BiExportRequest"
      responses:
        "202":
          description: Exportação enfileirada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BiExportJob"
        "400":
          description: Filtros inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Usuário não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /bi-exports/{exportId}:
    get:
      operationId: getBiExport
      tags: [engagement]
      summary: Consulta o andamento de uma exportação
      security:
        - sessionCookie: []
      parameters:
        - name: exportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Estado da exportação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BiExportJob"
        "401":
          description: Usuário não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Exportação não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /bi-exports/{exportId}/download:
    get:
      operationId: downloadBiExport
      tags: [engagement]
      summary: Baixa o arquivo CSV gerado
      security:
        - sessionCookie: []
      parameters:
        - name: exportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Arquivo CSV
          content:
            text/csv:
              schema:
                type: string
                format: binary
        "401":
          description: Usuário não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Arquivo inexistente ou expirado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: amoconecta_session
  schemas:
...
    SafetyMode:
      type: object
      properties:
        envio_liberado:
          type: boolean
        allowlist_count:
          type: integer
          minimum: 0
        message:
          type: ["string", "null"]
      required:
        - envio_liberado
        - allowlist_count
        - message
    CampaignDefaults:
      type: object
      properties:
        remetente_email:
          type: string
          format: email
        remetente_nome:
          type: string
        reply_to:
          type: ["string", "null"]
          format: email
        teto_hora:
          type: integer
          minimum: 0
        teto_dia:
          type: integer
          minimum: 0
        cor_botao_email:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        desengajados_total:
          type: integer
          minimum: 0
      required:
        - remetente_email
        - remetente_nome
        - reply_to
        - teto_hora
        - teto_dia
        - cor_botao_email
        - desengajados_total
    EmailBranding:
      type: object
      properties:
        cor_botao_email:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        atualizado_em:
          type: string
          format: date-time
      required: [cor_botao_email, atualizado_em]
    EmailBrandingUpdate:
      type: object
      properties:
        cor_botao_email:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
      required: [cor_botao_email]
    EngagementSummary:
      type: object
      properties:
        desengajados_total:
          type: integer
          minimum: 0
        calculado_em:
          type: ["string", "null"]
          format: date-time
      required: [desengajados_total, calculado_em]
    BiExportFilter:
      type: string
      enum: [todos, clicaram, abriram_sem_clicar, nao_abriram, bounce_ou_reclamacao]
    BiExportRequest:
      type: object
      properties:
        campanha_id:
          type: ["string", "null"]
          format: uuid
        periodo_inicio:
          type: ["string", "null"]
          format: date
        periodo_fim:
          type: ["string", "null"]
          format: date
        filtro:
          $ref: "#/components/schemas/BiExportFilter"
      required: [campanha_id, periodo_inicio, periodo_fim, filtro]
    BiExportJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        campanha_id:
          type: ["string", "null"]
          format: uuid
        filtro:
          $ref: "#/components/schemas/BiExportFilter"
        periodo_inicio:
          type: ["string", "null"]
          format: date
        periodo_fim:
          type: ["string", "null"]
          format: date
        status:
          type: string
          enum: [pendente, processando, concluida, erro, expirada]
        total_linhas:
          type: ["integer", "null"]
          minimum: 0
        linhas_processadas:
          type: integer
          minimum: 0
        criado_em:
          type: string
          format: date-time
        concluido_em:
          type: ["string", "null"]
          format: date-time
        expira_em:
          type: ["string", "null"]
          format: date-time
        erro:
          type: ["string", "null"]
        disponivel_para_download:
          type: boolean
      required:
        - id
        - campanha_id
        - filtro
        - periodo_inicio
        - periodo_fim
        - status
        - total_linhas
        - linhas_processadas
        - criado_em
        - concluido_em
        - expira_em
        - erro
        - disponivel_para_download
    BiExportJobsResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/BiExportJob"
      required: [jobs]
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...