# lib/api-zod/src/generated/types/biExportFilter.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API do AmoConecta
 * OpenAPI spec version: 0.1.0
 */

export type BiExportFilter = typeof BiExportFilter[keyof typeof BiExportFilter];


export const BiExportFilter = {
  todos: 'todos',
  clicaram: 'clicaram',
  abriram_sem_clicar: 'abriram_sem_clicar',
  nao_abriram: 'nao_abriram',
  bounce_ou_reclamacao: 'bounce_ou_reclamacao',
} as const;