# lib/api-zod/src/generated/types/biExportJob.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API do AmoConecta
 * OpenAPI spec version: 0.1.0
 */
import type { BiExportFilter } from './biExportFilter';
import type { BiExportJobStatus } from './biExportJobStatus';

export interface BiExportJob {
  id: string;
  /** @nullable */
  campanha_id: string | null;
  filtro: BiExportFilter;
  /** @nullable */
  periodo_inicio: Date | null;
  /** @nullable */
  periodo_fim: Date | null;
  status: BiExportJobStatus;
  /**
     * @minimum 0
     * @nullable
     */
  total_linhas: number | null;
  /** @minimum 0 */
  linhas_processadas: number;
  criado_em: Date;
  /** @nullable */
  concluido_em: Date | null;
  /** @nullable */
  expira_em: Date | null;
  /** @nullable */
  erro: string | null;
  disponivel_para_download: boolean;
}