# lib/api-zod/src/generated/types/biExportJobsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API do AmoConecta
 * OpenAPI spec version: 0.1.0
 */
import type { BiExportJob } from './biExportJob';

export interface BiExportJobsResponse {
  jobs: BiExportJob[];
}