# lib/api-zod/src/generated/types/biExportRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API do AmoConecta
 * OpenAPI spec version: 0.1.0
 */
import type { BiExportFilter } from './biExportFilter';

export interface BiExportRequest {
  /** @nullable */
  campanha_id: string | null;
  /** @nullable */
  periodo_inicio: Date | null;
  /** @nullable */
  periodo_fim: Date | null;
  filtro: BiExportFilter;
}