# lib/api-zod/src/generated/types/campaignDefaults.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API do AmoConecta
 * OpenAPI spec version: 0.1.0
 */

export interface CampaignDefaults {
  remetente_email: string;
  remetente_nome: string;
  /** @nullable */
  reply_to: string | null;
  /** @minimum 0 */
  teto_hora: number;
  /** @minimum 0 */
  teto_dia: number;
  /** @pattern ^#[0-9a-fA-F]{6}$ */
  cor_botao_email: string;
  /** @minimum 0 */
  desengajados_total: number;
}