# lib/api-zod/src/generated/types/emailBranding.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API do AmoConecta
 * OpenAPI spec version: 0.1.0
 */

export interface EmailBranding {
  /** @pattern ^#[0-9a-fA-F]{6}$ */
  cor_botao_email: string;
  atualizado_em: Date;
}