# lib/api-zod/src/generated/types/engagementSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API do AmoConecta
 * OpenAPI spec version: 0.1.0
 */

export interface EngagementSummary {
  /** @minimum 0 */
  desengajados_total: number;
  /** @nullable */
  calculado_em: Date | null;
}